Move main-page* into pages/
