+++ web/src/app/app.module.ts
@@ -26,7 +26,7 @@ import { USE_EMULATOR as USE_FUNCTIONS_EMULATOR } from '@angular/fire/compat/fun
 import { AppRoutingModule } from 'app/routing.module';
 import { AppComponent } from 'app/app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MainPageContainerModule } from 'app/components/main-page-container/main-page-container.module';
+import { MainPageContainerModule } from 'app/pages/main-page-container/main-page-container.module';
 import { environment } from 'environments/environment';
 import { JobDialogModule } from 'app/components/job-dialog/job-dialog.module';
 import { GoogleAuthProvider } from 'firebase/auth';
