+++ web/src/app/pages/main-page-container/main-page-container.component.css
