+++ web/src/app/pages/main-page-container/main-page-container.module.ts
@@ -16,8 +16,8 @@
 
 import { CommonModule } from '@angular/common';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { MainPageContainerComponent } from 'app/components/main-page-container/main-page-container.component';
-import { MainPageModule } from 'app/components/main-page/main-page.module';
+import { MainPageContainerComponent } from 'app/pages/main-page-container/main-page-container.component';
+import { MainPageModule } from 'app/pages/main-page-container/main-page/main-page.module';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
