+++ web/src/app/pages/main-page-container/main-page/main-page.component.spec.ts
@@ -16,7 +16,7 @@
 
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { Component, NO_ERRORS_SCHEMA } from '@angular/core';
-import { MainPageComponent } from 'app/components/main-page/main-page.component';
+import { MainPageComponent } from 'app/pages/main-page-container/main-page/main-page.component';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ActivatedRouteStub } from 'testing/activated-route-stub';
 import { SurveyService } from 'app/services/survey/survey.service';
