+++ web/src/app/pages/main-page-container/main-page/main-page.module.ts
@@ -18,7 +18,7 @@ import { CommonModule } from '@angular/common';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { JobDialogModule } from 'app/components/job-dialog/job-dialog.module';
 import { DrawingToolsModule } from 'app/components/drawing-tools/drawing-tools.module';
-import { MainPageComponent } from 'app/components/main-page/main-page.component';
+import { MainPageComponent } from 'app/pages/main-page-container/main-page/main-page.component';
 import { MapModule } from 'app/components/map/map.module';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialogModule } from '@angular/material/dialog';
@@ -26,7 +26,7 @@ import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { NgModule } from '@angular/core';
 import { SidePanelModule } from 'app/components/side-panel/side-panel.module';
-import { SurveyHeaderModule } from 'app/components/main-page/survey-header/survey-header.module';
+import { SurveyHeaderModule } from 'app/pages/main-page-container/main-page/survey-header/survey-header.module';
 import { MatIconModule } from '@angular/material/icon';
 
 @NgModule({
