+++ web/src/app/pages/main-page-container/main-page/survey-header/survey-header.component.scss
