+++ web/src/app/pages/main-page-container/main-page/survey-header/survey-header.component.spec.ts
@@ -16,7 +16,7 @@
 
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { InlineEditorModule } from 'app/components/inline-editor/inline-editor.module';
-import { SurveyHeaderComponent } from 'app/components/main-page/survey-header/survey-header.component';
+import { SurveyHeaderComponent } from 'app/pages/main-page-container/main-page/survey-header/survey-header.component';
 import { SurveyService } from 'app/services/survey/survey.service';
 import { MatIconModule } from '@angular/material/icon';
 import { MatDialogModule } from '@angular/material/dialog';
